adicionando funcao para mudar o valor das variaveis de ambiente OLD_PWD e PWD, quando se executa o comando cd
